BackEnd/AI_BackEnd_django/requirements.txt
djangorestframework
numpy
djongo
keras
scikit-learn
channels
django-cors-headers
python-decouple
mysqlclient
opencv-python
tensorflow
keras-facenet